ai/requirements.txt
# Web Framework
fastapi>=0.110
uvicorn>=0.29
uvicorn==0.34.0
pydantic==2.10.6

# python-multipart==0.0.6
# env
python-dotenv==1.0.0

# Server Connection
requests>=2.31 # HTTP requests
pyyaml>=6.0 # YAML parsing
loguru>=0.6.0 # Logging

# -------------------------------------------------------------------

#### Food Recog ####

# Deep Learning Framework
torch>=2.1
torchvision>=0.16

# Object Detection (YOLOv8)
ultralytics>=8.0.166

# Computer Vision
opencv-python>=4.8
opencv-contrib-python>=4.8 # ArUco marker, Stereo Vision
# opencv-python-headless>=4.8 # For Docker

# DB connection
boto3>=1.34.0 # S3
aiobotocore>=2.7 # optional, for async S3 access

# Monocular Depth Estimation(MiDas/DPT)
timm>=0.9.12 # timm models

# data processing
numpy>=1.24
pandas>=2.0


# -------------------------------------------------------------------

#### AGENT LLM ####

# Language Model Framework
langchain>=0.0.268
langchain-community>=0.0.20
langgraph>=0.1.0

# LLM API Clients
openai>=0.27.6
anthropic>=0.4.0

# DB connection
SQLAlchemy>=2.0.20
PyMySQL>=1.0.3
## PostgreSQL + pgvector
# psycopg2-binary>=2.9.6
# pgvector>=0.7.0


# -------------------------------------------------------------------
